Refactor because survey routes are getting a lot
