+++ src/routes/mod.rs
@@ -1,4 +1,2 @@
 pub(crate) mod index;
-pub(crate) mod net_promoter_score;
-pub(crate) mod attrakdiff;
-pub(crate) mod system_usability_score;
+pub(crate) mod survey;
